Refactor PathSettings and TestKinds
