+++ src/main/scala/scala/tools/partest/nest/Runner.scala
@@ -344,7 +344,7 @@ class Runner(val testFile: File, val suiteRunner: AbstractRunner, val nestUI: Ne
     )
 
     def masters    = {
-      val files = List(new File(parentFile, "filters"), new File(PathSettings.srcDir.path, "filters"))
+      val files = List(new File(parentFile, "filters"), new File(suiteRunner.pathSettings.srcDir.path, "filters"))
       files filter (_.exists) flatMap (_.fileLines) map (_.trim) filter (s => !(s startsWith "#"))
     }
     val filters    = toolArgs("filter", split = false) ++ masters
@@ -525,7 +525,7 @@ class Runner(val testFile: File, val suiteRunner: AbstractRunner, val nestUI: Ne
   }
 
   def extraClasspath = kind match {
-    case "specialized"  => List(PathSettings.srcSpecLib.fold(sys.error, identity))
+    case "specialized"  => List(suiteRunner.pathSettings.srcSpecLib.fold(sys.error, identity))
     case _              => Nil
   }
   def extraJavaOptions = kind match {
